pfbapi/requirements.txt
aiohttp==3.7.4.post0
aniso8601==9.0.1
async-timeout==3.0.1
asyncio==3.4.3
attrs==20.3.0
backoff==1.10.0
boto3==1.17.35
botocore==1.20.35
cdislogging==1.0.0
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
dictionaryutils
drsclient
fastavro==1.3.3
Flask==1.1.2
Flask-Cors==3.0.10
Flask-Jsonpify==1.5.0
Flask-RESTful==0.3.8
Flask-SQLAlchemy==2.4.4
gdcdictionary==1.2.0
gen3==4.2.0
greenlet==1.0.0
h11==0.9.0
httpcore==0.11.1
httpx==0.15.5
idna==2.10
importlib-metadata==1.7.0
indexclient==2.1.0
itsdangerous==1.1.0
Jinja2==2.11.3
jmespath==0.10.0
jsonschema
MarkupSafe==1.1.1
multidict==5.1.0
numpy==1.20.1
pandas==1.2.3
pypfb==0.5.8
pyrsistent==0.17.3
python-dateutil==2.8.1
python-json-logger==0.1.11
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
rfc3986==1.4.0
s3transfer==0.3.6
six==1.15.0
sniffio==1.2.0
SQLAlchemy==1.4.1
typing-extensions==3.7.4.3
urllib3==1.26.4
Werkzeug==1.0.1
yarl==1.6.3
zipp==3.4.1
